requirements.txt
SQLAlchemy==2.0.3
python-dotenv==0.21.1
psycopg2==2.9.5